pyx compiling, testing framework in place

+++ gpvolve/markov/utils/_generate_tmatrix/fixation.h
@@ -1,8 +1,8 @@
 #ifndef FIXATION_H
 #define FIXATION_H
 
-double moran(double fitness_i, double fitness_j, int population_size);
-double mcclandish(double fitness_i, double fitness_j, int population_size);
-double sswm(double fitness_i, double fitness_j, int population_size);
+double moran(double fitness_i, double fitness_j, long population_size);
+double mcclandish(double fitness_i, double fitness_j, long population_size);
+double sswm(double fitness_i, double fitness_j, long population_size);
 
 #endif /* FIXATION_H */
